+++ app/src/main/java/com/yuhang/novel/pirate/repository/network/data/kanshu/result/RecommendBooksResult.kt
@@ -4,6 +4,6 @@ data class RecommendBooksResult(val Img: String = "",
                                  val Score: Int = 0,
                                  val LastChapter: String = "",
                                  val Author: String = "",
-                                 val Id: Int = 0,
+                                 val Id: Long = 0,
                                  val LastChapterId: Int = 0,
                                  val Name: String = "")
