+++ app/src/main/java/com/yuhang/novel/pirate/service/UsersService.kt
@@ -11,18 +11,18 @@ interface UsersService {
     /**
      * 更新线上收藏记录到本地
      */
-    fun updateCollectionToLocal(): Flowable<Int>
+    fun updateCollectionToLocal(): Flowable<Long>
 
 
     /**
      * 更新小说章节列表到本地
      */
-    fun updateChapterListToLocal(bookid: Int): Flowable<Int>
+    fun updateChapterListToLocal(bookid: Long): Flowable<Long>
 
     /**
      * 更新小说详情到本地
      */
-    fun updateBookInfoToLocal(bookid: Int): Flowable<Int>
+    fun updateBookInfoToLocal(bookid: Long): Flowable<Long>
 
     /**
      * 更新用户信息到本地
@@ -32,5 +32,5 @@ interface UsersService {
     /**
      * 更新内容到本地,更新第一章内容
      */
-    fun updateContentToLocal(bookid: Int):Flowable<Int>
+    fun updateContentToLocal(bookid: Long):Flowable<Long>
 }
